requirements.txt
flask
pillow
requests
gunicorn
waitress
